+++ R/interpolate.R
@@ -1,5 +1,26 @@
-#' @name interpolate_signal
-#' @title Interpolate the missing points and unify sampling interval for the input sensor data
+#' Interpolate missing points and unify sampling rate for multi-channel signal
+#'
+#' \code{interpolate_signal} applies different interpolation algorithms to the
+#' input multi-channel signal to fill in the missing samples and harmonizes the
+#' sampling rate.
+#'
+#' @section How is it used in MIMS-unit algorithm?: This function is a utility
+#'   function that has been used in functions: \code{\link{extrapolate}}, and
+#'   \code{\link{simulate_new_data}}.
+#'
+#' @param df dataframe. Input multi-channel accelerometer signal.
+#' @param method string. Interpolation algorithms. Could be "spline_natural",
+#'   "spline_improved" or "spline_fmm": see \code{\link[stats]{spline}};
+#'   "aspline_original": see \code{\link[akima]{spline}}; and "linear": see
+#'   \code{\link[stats]{approx}}. Default is "spline_natural".
+#' @param sr number. Sampling rate in Hz of the input signal. Default is 100.
+#' @param st POSIXct date. The start time for interpolation. If it is
+#'   \code{NULL}, it will use the start time of the input signal. Default is
+#'   \code{NULL}.
+#' @param et POSIXct date. The end time for interpolation. If it is \code{NULL},
+#'   it will use the end time of the input signal. Default is \code{NULL}.
+#' @return dataframe. Interpolated signal.
+#' @family utility functions
 #' @export
 interpolate_signal <-
   function(df,
@@ -48,7 +69,7 @@ interpolate_signal <-
             xout = x_out,
             method = "original"
           ),
-          aspline_improved = stats::spline(
+          spline_improved = stats::spline(
             x = ts,
             y = col_data,
             xout = x_out,