+++ R/mims_unit_.R
@@ -35,7 +35,7 @@ mims_unit_ <- function(df, breaks = "5 sec", range, noise_level = 0.03, k = 0.05
     # apply extrapolation algorithm
     if (use_extrapolation)
     {
-        extrapolatedData <- extrapolate.data.frame(df, range, noise_level, k, spar)
+        extrapolatedData <- extrapolate(df, range, noise_level, k, spar)
     } else if (use_interpolation)
     {
         extrapolatedData <- interpolate_signal(df, sr = 100, method = "spline_natural")