+++ data-raw/treadmill_make_data.R
@@ -10,7 +10,7 @@ files = list.files(path = folder, pattern = ".*RAW\\.csv", full.names = TRUE, re
 
 l_ply(files, function(sensor_file){
   csvData = import_actigraph_csv(sensor_file)
-  data2g = MIMSunit::make_sensor_data(oldData = csvData, new_range = c(-2, 2), new_sr = sampling_rate(csvData))
+  data2g = MIMSunit::simulate_new_data(oldData = csvData, new_range = c(-2, 2), new_sr = sampling_rate(csvData))
   output_file = str_replace(sensor_file, "RAW\\.csv", "RAW_2g\\.sensor\\.csv")
   options(digits.secs=3)
   options(digits = 6)